requirements.txt
Flask>=2.0
line-bot-sdk>=2.0.0
python-dotenv>=0.20.0
